Module aws_smithy_eventstream::frame

Event Stream message frame types and serialization/deserialization logic.

Structs

Header

Event Stream header.

Message

Event Stream message.

MessageFrameDecoder

Streaming decoder for decoding a Message from a stream.

Enums

DecodedFrame

Return value from MessageFrameDecoder.

HeaderValue

Event Stream frame header value.

UnmarshalledMessage

A successfully unmarshalled message that is either an Event or an Error.

Traits

MarshallMessage

Converts a Smithy modeled Event Stream type into a Message.

SignMessage

Signs an Event Stream message.

UnmarshallMessage

Converts an Event Stream Message into a Smithy modeled type.

Functions

write_headers_to

Writes the given headers to a buffer.

Type Definitions

SignMessageError