aws_smithy_eventstream

Module frame

Event Stream message frame types and serialization/deserialization logic.

Structs

DeferredSigner

Deferred event stream signer to allow a signer to be wired up later.

DeferredSignerSender

A sender that gets placed in the request config to wire up an event stream signer after signing.

MessageFrameDecoder

Streaming decoder for decoding a Message from a stream.

NoOpSigner

Enums

DecodedFrame

Return value from MessageFrameDecoder.

UnmarshalledMessage

A successfully unmarshalled message that is either an Event or an Error.

Traits

MarshallMessage

Converts a Smithy modeled Event Stream type into a Message.

SignMessage

Signs an Event Stream message.

UnmarshallMessage

Converts an Event Stream Message into a Smithy modeled type.

Functions

read_message_from

Reads a message from the given buffer. For streaming use cases, use the MessageFrameDecoder instead of this.

write_headers_to

Writes the given headers to a buffer.

write_message_to

Writes the message to the given buffer.

Type Aliases

SignMessageError