aws_smithy_eventstream::frame

Trait SignMessage

pub trait SignMessage: Debug {
    // Required methods
    fn sign(&mut self, message: Message) -> Result<Message, SignMessageError>;
    fn sign_empty(&mut self) -> Option<Result<Message, SignMessageError>>;
}

Signs an Event Stream message.

Required Methods

fn sign(&mut self, message: Message) -> Result<Message, SignMessageError>

fn sign_empty(&mut self) -> Option<Result<Message, SignMessageError>>

SigV4 requires an empty last signed message to be sent. Other protocols do not require one. Return Some(_) to send a signed last empty message, before completing the stream. Return None to not send one and terminate the stream immediately.